docker/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
APScheduler==3.6.3
attrs==23.1.0
cachetools==4.2.2
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
magic-filter==1.0.12
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tornado==6.1
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
yarl==1.9.3
